Anspass: Initial push.
This is the initial push of the anspass/anspassd/anspass-ctrl utilities.
This set of utilities can be used to automatically answer queries by 3rd
party applications that implement the askpass interface such as git.
See git credentials (https://git-scm.com/docs/gitcredentials) for more
details about the interface.

The correct environment variables need to be configured and the
credentials need to be added prior to operation.  Please see the
README.txt for more details.

Signed-off-by: Liam R. Howlett <Liam.Howlett@WindRiver.com>